refactor: restructured code delete: removed zod-validation middleware use async handler for checking request validity
