+++ src/routes/index.ts
@@ -4,6 +4,8 @@ import manageRouter from "./rides/manage/index.ts";
 import searchRouter from "./rides/search/index.ts";
 import ridesRouter from "./rides/rides.ts";
 
+import userRouter from "./user/user.ts";
+
 const router = express.Router();
 
 router.use("/user", userRouter);
