--- test/postreq.json
@@ -1,4 +1,0 @@
-{
-  "name": "johndoe",
-  "email": "exdelm"
-}
